requirements.txt
python-telegram-bot
nltk
scikit-learn